Create file for recipes and link to main page
Recipes will be stored in their own folder and all links
will go to this folder. The first recipe is for butternut squash
so the page is set up to begin writing down that recipe.

The index page is also updated to use the link and seems to work
properly.