Add stylesheet and apply style to home page
Now that I have some abilities in styling, I should apply basic
styles to the pages of the website. I linked a single style sheet to all
pages so that there is a uniform look. The only page with real changes
is the home page and the classes were added to that page to get things
working.

Note that the h1 tag is on each page so the style is applied to all
of the page titles.